+++ CHANGELOG.md
@@ -32,10 +32,17 @@
 * Modified credits scene
   * Added `typographics` to the in-game credits, for making the `Coin Bag` item sprite.
   * Removed `Jdaster64` from the in-game credits, as he was originally added there due to a misunderstanding. Sorry Jdaster, hope you don't mind too much :)
+* Changed the internal handling of `Require Specific Spirits` and `Limit Chapter Logic`. These settings are now merged into a single setting called `RequiredSpirits`.
 * RAM locations docs
   * Fix wrong item check collected flag for Boo's Mansion - Super Boots Room - In Big Chest. This should stop auto-trackers marking the item as collected even if the item is never picked up.
 * New bomb trigger for Toad Town playroom
   * We originally moved the Bombette bomb trigger for opening the Toad Town playroom over to the tree, because the bomb trigger was on the wrong side of the map. We now moved that bomb trigger back to its original, wrong position, and added a second bomb trigger next to the tree. This way both of these locations work for opening the playroom, so you can show off knowing this obscure bug to your friends again :)
+
+### Additional Technical Changelog
+
+* Yaml settings changes:
+  * Renamed `RequireSpecificSpirits` to `RequiredSpirits`, and is now int (was bool)
+  * Removed `LimitChapterLogic`
 
 ## 0.29.1 (beta)